refactor(server/shared): reduce (but not fix) circular dependencies
