+++ packages/server/modules/shared/roles.js
@@ -1,0 +1,12 @@
+const knex = require(`@/db/knex`)
+let roles
+
+const getRoles = async () => {
+  if (roles) return roles
+  roles = await knex('user_roles').select('*')
+  return roles
+}
+
+module.exports = {
+  getRoles
+}
